+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - Use ggwsm in pipelines
+- Install the dna_ggsuite DNA
 
 ## 3.2.0 - 2026-08-14